+++ source/Stairs.h
@@ -5,19 +5,21 @@
 
 
 
-class cStairs												// tolua_export
-{															// tolua_export
+// tolua_begin
+
+class cStairs
+{
 public:
 	/// Converts player rotation to stair rotation metadata. To get upside-down stairs, OR with 0x4
-	static NIBBLETYPE RotationToMetaData(float a_Rotation)  // tolua_export
-	{														// tolua_export
-    a_Rotation += 90 + 45; // So its not aligned with axis
-    NIBBLETYPE result = 0x0;
-		if (a_Rotation > 360.f)
+	static NIBBLETYPE RotationToMetaData(double a_Rotation)
+	{
+		a_Rotation += 90 + 45; // So its not aligned with axis
+		NIBBLETYPE result = 0x0;
+		if (a_Rotation > 360)
 		{
-			a_Rotation -= 360.f;
+			a_Rotation -= 360;
 		}
-		if ((a_Rotation >= 0.f) && (a_Rotation < 90.f))
+		if ((a_Rotation >= 0) && (a_Rotation < 90))
 		{
 			return 0x0;
 		}
@@ -33,9 +35,10 @@ public:
 		{
 			return 0x3;
 		}
-	}														// tolua_export
-} ;															// tolua_export
+	}
+} ;
 
+// tolua_end
