+++ source/UI/SlotArea.cpp
@@ -764,7 +764,7 @@ void cSlotAreaTemporary::TossItems(cPlayer & a_Player, int a_Begin, int a_End)
 		Item.Empty();
 	}  // for i - itr->second[]
 	
-	float vX = 0, vY = 0, vZ = 0;
+	double vX = 0, vY = 0, vZ = 0;
 	EulerToVector(-a_Player.GetRotation(), a_Player.GetPitch(), vZ, vX, vY);
 	vY = -vY * 2 + 1.f;
 	a_Player.GetWorld()->SpawnItemPickups(Drops, a_Player.GetPosX(), a_Player.GetPosY() + 1.6f, a_Player.GetPosZ(), vX * 2, vY * 2, vZ * 2);
